+++ src/lib.rs
@@ -117,6 +117,8 @@
 mod error;
 mod render;
 mod tile;
+// TODO: if possible, try to reuse the implementation from h3o instead.
+mod ring_hierarchy;
 
 pub use error::RenderingError;
 pub use render::render;